+++ eventstore/api/reactor/src/main/java/org/occurrent/eventstore/api/reactor/UnconditionallyWriteToEventStream.java
@@ -21,10 +21,24 @@ import org.occurrent.eventstore.api.WriteResult;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.util.Objects;
+
 /**
  * An interface that should be implemented by event streams that supports writing events to a stream without specifying a write condition.
  */
 public interface UnconditionallyWriteToEventStream {
+
+    /**
+     * Write a single {@code event} to a stream
+     *
+     * @param streamId The stream id of the stream to write to
+     * @param event    The events to write
+     */
+    default Mono<WriteResult> write(String streamId, CloudEvent event) {
+        Objects.requireNonNull(event, CloudEvent.class.getSimpleName() + " cannot be null");
+        return write(streamId, Flux.just(event));
+    }
+
     /**
      * Write {@code events} to a stream
      *